Moved msg folder. Add msgs for example app2
