+++ examples/Messaging/App2/MsgProcessor.hpp
@@ -5,5 +5,12 @@
 
 namespace app2
 {
+    class MsgHandler
+    {
+        public:
+            MsgHandler()
+            {}
 
+            
+    };
 }
