+++ inc/etfw/msg/MsgHelper.hpp
